refactor(ui): move quick-filter shared types below the component layer
The PR quality gate flagged layering warnings on the files this branch
touches: AdvancedSearchPureUtils (a pure util) imported types and an enum
from component interfaces, and useQuickFilterLabels (a hook) imported a
component type.

- SearchDropdownOption and ExploreQuickFilterField move to
  src/interface/quickFilter.interface.ts; AssetsOfEntity moves to
  src/enums/Assets.enum.ts. The original modules import and re-export them,
  so the other ~30 importers are untouched.
- Page/type imports that only carry types are now import type
  (QueryFilterInterface, ExploreSearchIndex).

Remaining warnings on ExploreQuickFilters.tsx are pre-existing runtime
cycles (SearchClassBase, SearchDropdown, AdvanceSearchProvider) and the
ES-query type family living under pages/ — the SearchDropdown edge goes
away with the FilterSelect swap, the rest need their own refactor.
