+++ openmetadata-ui/src/main/resources/ui/src/components/Glossary/GlossaryTerms/tabs/AssetsTabs.interface.ts
@@ -15,20 +15,9 @@ import { OperationPermission } from '../../../../context/PermissionProvider/Perm
 import { EntityDetailsObjectInterface } from '../../../Explore/ExplorePage.interface';
 import { SearchedDataProps } from '../../../SearchedData/SearchedData.interface';
 
-export enum AssetsOfEntity {
-  GLOSSARY = 'GLOSSARY',
-  DOMAIN = 'DOMAIN',
-  DATA_PRODUCT = 'DATA_PRODUCT',
-  DATA_PRODUCT_INPUT_PORT = 'DATA_PRODUCT_INPUT_PORT',
-  DATA_PRODUCT_OUTPUT_PORT = 'DATA_PRODUCT_OUTPUT_PORT',
-  TEAM = 'TEAM',
-  MY_DATA = 'MY_DATA',
-  FOLLOWING = 'FOLLOWING',
-  ACCESS_TOKEN = 'ACCESS_TOKEN',
-  LINEAGE = 'LINEAGE',
-  TAG = 'TAG',
-  COLUMN = 'COLUMN',
-}
+import { AssetsOfEntity } from '../../../../enums/Assets.enum';
+
+export { AssetsOfEntity };
 
 export interface AssetsTabsProps {
   onAddAsset: () => void;
