+++ openmetadata-ui/src/main/resources/ui/src/components/SearchDropdown/SearchDropdown.interface.ts
@@ -11,8 +11,7 @@
  *  limitations under the License.
  */
 
-import { ReactNode } from 'react';
-import { ExploreSearchIndex } from '../Explore/ExplorePage.interface';
+import type { ExploreSearchIndex } from '../Explore/ExplorePage.interface';
 
 export interface SearchDropdownProps {
   triggerButtonSize?: 'large' | 'middle' | 'small';
@@ -44,11 +43,6 @@ export interface SearchDropdownProps {
   getPopupContainer?: (triggerNode: HTMLElement) => HTMLElement;
 }
 
-export interface SearchDropdownOption {
-  key: string;
-  label: string;
-  labelKeyOptions?: Record<string, string | number | boolean>;
-  count?: number;
-  description?: string;
-  icon?: ReactNode;
-}
+import type { SearchDropdownOption } from '../../interface/quickFilter.interface';
+
+export type { SearchDropdownOption };
